Create Hangman game
Here's a C++ console application that simulates a game of Hangman. The program randomly selects a word from a list of words and allows the user to guess letters one at a time. If the user correctly guesses a letter, the program reveals the location of that letter in the word. If the user makes too many incorrect guesses, the game ends, and the correct word is revealed. We'll use a std::string to store the selected word and the user's guesses: